Set user runtime path dynamically for Vundle.
The default for Vundle is to look for the bundle directory in ~/.vim. If
you're trying to use your user runtime directory with Windows, too,
that's a problem since the directory is then vimfiles. One solution, as
implemented here, is to dynamically determine the directory from the
location of vimrc. It works if you keep your vimrc in the user runtime
directory and as long as you don't have a stray one in your home
directory. It seems at least some versions of Vim for Windows will add
.vim in the home directory to the runtime path which makes
vundle#begin() work. The problem is that vim doesn't seem to look for
vimrc and gvimrc in .vim under those circumstances and you have to do
something like adding symlinks in home.